# packages/cli-init/templates/frontend/src/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Jumentix API",
    "description": "`REST API` built with `Jumentix` and `Domain Designer`",
    "termsOfService": "https://www.xpertminds.dev/terms/",
    "contact": {
      "email": "apiteam@xpertminds.dev"
    },
    "license": {
      "name": "Unlicensed",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0",
    "x-rbac": {
      "description": "Role-to-scope matrix enforced by the backend and mirrored here for OAS-driven clients.",
      "roles": {
        "superadmin": ["*"],
        "admin": [
          "access_allow",
          "read_user",
          "create_user",
          "update_user",
          "delete_user",
          "read_organization",
          "update_organization"
        ],
        "user": ["access_allow", "read_user"]
      }
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api/1.0.0",
      "x-service-id": "core"
    }
  ],
...
  "paths": {
...
    "/users/{id}": {
      "delete": {
        "tags": ["user"],
        "summary": "Delete user",
        "description": "Deletes a single user",
        "operationId": "deleteOne",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to return",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceDeleteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearerAuth": ["delete_user"]
          }
        ]
      },
      "put": {
        "tags": ["user"],
        "summary": "Update user",
        "description": "Updates a single user",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Update an user from Store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestUpdateUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "security": [
          {
            "bearerAuth": ["update_user"]
          }
        ]
      },
      "get": {
        "tags": ["user"],
        "summary": "Get an user by ID",
        "description": "Returns a single user",
        "operationId": "getOneById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to return",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearerAuth": ["read_user"]
          }
        ]
      }
    },
...
    "/organizations": {
      "get": {
        "tags": ["organization"],
        "summary": "Get all organizations",
        "description": "Returns a page of organizations. Delta sync page: sort=`updatedAt:asc,id:asc`, filter `{\"updatedAt\":{\"operator\":\"gt\",\"value\":\"<iso>\"}}`, page through `page`/`size`, set `includeDeleted=true` so tombstones travel with the cursor (`x-sync`).",
        "operationId": "getAllOrganizations",
        "x-service": "core",
        "x-sync": {
          "cursorField": "updatedAt",
          "includeDeletedParam": "includeDeleted"
        },
        "x-list-capabilities": {
          "sortable": ["name", "createdAt", "updatedAt", "id"],
          "filterable": {
            "name": "text",
            "createdAt": "date",
            "updatedAt": "date"
          },
          "searchable": ["name"],
          "defaultSize": 30,
          "maxSize": 100,
          "includeDeleted": true
        },
        "x-metrics-capabilities": {
          "groupable": [],
          "series": ["createdAt", "updatedAt"]
        },
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number to fetch (1-based).",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Page size; bounded by `x-list-capabilities.maxSize`.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 30
            }
          },
          {
            "name": "filter",
            "in": "query",
            "description": "Base64-encoded JSON object keyed by a field listed in `x-list-capabilities.filterable`. A scalar value means equality; an object `{ \"operator\": \"contains\" | \"eq\" | \"ne\" | \"in\" | \"gte\" | \"lte\" | \"between\", \"value\": ... }` applies that operator. Unknown fields or operators are rejected with 400.",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 4096
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Comma-separated `field:asc|desc` pairs, fields limited to `x-list-capabilities.sortable`. Unknown fields are rejected with 400.",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 256,
              "pattern": "^[A-Za-z0-9_.]+(:(asc|desc))?(,[A-Za-z0-9_.]+(:(asc|desc))?)*$"
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Case-insensitive free-text search over `x-list-capabilities.searchable`.",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "includeDeleted",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationArrayOf"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearerAuth": ["read_organization"]
          }
        ]
      },
      "post": {
        "tags": ["organization"],
        "summary": "Add a new organization",
        "description": "Add a new organization",
        "operationId": "createOrganization",
        "requestBody": {
          "description": "Create organization payload",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCreateOrganization"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Organization created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "security": [
          {
            "bearerAuth": ["create_organization"]
          }
        ]
      }
    },
...
    "/organizations/{id}": {
      "get": {
        "tags": ["organization"],
        "summary": "Get organization by ID",
        "description": "Returns a single organization",
        "operationId": "getOrganizationById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Organization ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Organization not found"
          }
        },
        "security": [
          {
            "bearerAuth": ["read_organization"]
          }
        ]
      },
      "put": {
        "tags": ["organization"],
        "summary": "Update organization",
        "description": "Updates a single organization",
        "operationId": "updateOrganization",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Organization ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Organization update payload",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestUpdateOrganization"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Organization not found"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "security": [
          {
            "bearerAuth": ["update_organization"]
          }
        ]
      },
      "delete": {
        "tags": ["organization"],
        "summary": "Delete organization",
        "description": "Deletes a single organization",
        "operationId": "deleteOrganization",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Organization ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceDeleteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Organization not found"
          }
        },
        "security": [
          {
            "bearerAuth": ["delete_organization"]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ResourceDeleteResponse": {
        "description": "Port output object for delete operations.",
        "required": ["data"],
        "type": "object",
        "properties": {
          "data": {
            "x-label": {
              "en": "Number",
              "pt-BR": "Número"
            },
            "type": "boolean",
            "default": false,
            "description": "Result of request to delete resource"
          }
        }
      },
...
      "RequestUpdateUser": {
        "description": "Port input object for user update endpoint.",
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": {
            "x-label": {
              "en": "ID",
              "pt-BR": "ID"
            },
            "type": "string"
          },
          "firstName": {
            "x-label": {
              "en": "First name",
              "pt-BR": "Nome"
            },
            "type": "string",
            "minLength": 1,
            "example": "Abraham",
            "description": "User's first name"
          },
          "lastName": {
            "x-label": {
              "en": "Last name",
              "pt-BR": "Sobrenome"
            },
            "type": "string",
            "example": "Abraham",
            "nullable": true,
            "description": "User's last name"
          },
          "avatar": {
            "x-label": {
              "en": "Avatar",
              "pt-BR": "Avatar"
            },
            "type": "string",
            "maxLength": 255,
            "example": "avatar.png",
            "default": "avatar.png",
            "description": "User's avatar"
          },
          "username": {
            "x-label": {
              "en": "Username",
              "pt-BR": "Usuário"
            },
            "type": "string",
            "minLength": 1,
            "description": "Username to access the system"
          },
          "organization": {
            "x-label": {
              "en": "Organization",
              "pt-BR": "Organização"
            },
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Organization id for tenant users (admin/user)",
            "x-relation": {
              "entity": "Organization",
              "match": "id",
              "display": "name",
              "kind": "belongsTo"
            }
          },
          "emails": {
            "x-label": {
              "en": "Emails",
              "pt-BR": "E-mails"
            },
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Email"
            },
            "description": "User's e-mail addresses",
            "minItems": 1
          },
          "documents": {
            "x-label": {
              "en": "Documents",
              "pt-BR": "Documentos"
            },
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Document"
            },
            "description": "User's attached documents"
          },
          "phones": {
            "x-label": {
              "en": "Phones",
              "pt-BR": "Telefones"
            },
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Phone"
            },
            "description": "User's attached phone numbers"
          },
          "roles": {
            "x-label": {
              "en": "Roles",
              "pt-BR": "Papéis"
            },
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "User's access roles"
          }
        }
      },
      "RequestCreateOrganization": {
        "description": "Port input object for organization creation endpoint.",
        "required": ["name"],
        "type": "object",
        "properties": {
          "id": {
            "x-label": {
              "en": "ID",
              "pt-BR": "ID"
            },
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "x-label": {
              "en": "Name",
              "pt-BR": "Nome"
            },
            "type": "string",
            "minLength": 1
          },
          "address": {
            "x-label": {
              "en": "Addresses",
              "pt-BR": "Endereços"
            },
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Address"
            }
          },
          "phone": {
            "x-label": {
              "en": "Phones",
              "pt-BR": "Telefones"
            },
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Phone"
            }
          },
          "email": {
            "x-label": {
              "en": "Email",
              "pt-BR": "E-mail"
            },
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Email"
            }
          },
          "users": {
            "x-label": {
              "en": "Members",
              "pt-BR": "Membros"
            },
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "RequestUpdateOrganization": {
        "description": "Port input object for organization update endpoint.",
        "required": ["id"],
        "type": "object",
        "properties": {
          "id": {
            "x-label": {
              "en": "ID",
              "pt-BR": "ID"
            },
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "x-label": {
              "en": "Name",
              "pt-BR": "Nome"
            },
            "type": "string",
            "minLength": 1
          },
          "address": {
            "x-label": {
              "en": "Addresses",
              "pt-BR": "Endereços"
            },
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Address"
            }
          },
          "phone": {
            "x-label": {
              "en": "Phones",
              "pt-BR": "Telefones"
            },
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Phone"
            }
          },
          "email": {
            "x-label": {
              "en": "Email",
              "pt-BR": "E-mail"
            },
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Email"
            }
          },
          "users": {
            "x-label": {
              "en": "Members",
              "pt-BR": "Membros"
            },
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "OrganizationArrayOf": {
        "description": "Port output page of organizations: the `result` slice selected by `page`/`size`/`filter`/`sort`/`q` plus the matching `total` (JUM-777).",
        "type": "object",
        "required": ["result", "page", "size", "total"],
        "properties": {
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Organization"
            }
          },
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "size": {
            "type": "integer",
            "minimum": 1
          },
          "total": {
            "type": "integer",
            "minimum": 0,
            "description": "Number of records matching the filters and search, across all pages."
          }
        }
      },
...
      "Email": {
        "description": "Value object contract for Email.",
        "type": "object",
        "required": ["email", "type"],
        "properties": {
          "id": {
            "x-label": {
              "en": "ID",
              "pt-BR": "ID"
            },
            "type": "string"
          },
          "email": {
            "x-label": {
              "en": "Email",
              "pt-BR": "E-mail"
            },
            "type": "string",
            "format": "email",
            "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
            "maxLength": 254
          },
          "type": {
            "x-label": {
              "en": "Type",
              "pt-BR": "Tipo"
            },
            "type": "string",
            "enum": ["work", "personal"]
          },
          "isPrimary": {
            "x-label": {
              "en": "Primary",
              "pt-BR": "Principal"
            },
            "type": "boolean",
            "example": true,
            "description": "Is this the primary User's e-mail address?",
            "default": false
          }
        }
      },
      "Address": {
        "description": "Value object contract for Address.",
        "type": "object",
        "required": ["email", "type"],
        "properties": {
          "id": {
            "x-label": {
              "en": "ID",
              "pt-BR": "ID"
            },
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "x-label": {
              "en": "Email",
              "pt-BR": "E-mail"
            },
            "type": "string"
          },
          "type": {
            "x-label": {
              "en": "Type",
              "pt-BR": "Tipo"
            },
            "type": "string",
            "enum": ["work", "home", "vacation"]
          },
          "isPrimary": {
            "x-label": {
              "en": "Primary",
              "pt-BR": "Principal"
            },
            "type": "boolean",
            "default": false
          }
        }
      },
...
      "User": {
        "description": "Port output object for User resource.",
        "x-primary-key": "id",
        "x-service": "core",
        "required": ["id", "createdAt", "updatedAt", "firstName", "username", "password", "emails"],
        "type": "object",
        "properties": {
          "id": {
            "x-label": {
              "en": "ID",
              "pt-BR": "ID"
            },
            "type": "string",
            "format": "uuid"
          },
          "firstName": {
            "x-label": {
              "en": "First name",
              "pt-BR": "Nome"
            },
            "type": "string",
            "minLength": 1,
            "example": "Abraham",
            "description": "User's first name"
          },
          "lastName": {
            "x-label": {
              "en": "Last name",
              "pt-BR": "Sobrenome"
            },
            "type": "string",
            "example": "Abraham",
            "nullable": true,
            "description": "User's last name"
          },
          "username": {
            "x-label": {
              "en": "Username",
              "pt-BR": "Usuário"
            },
            "type": "string",
            "minLength": 1,
            "description": "Username to access the system"
          },
          "organization": {
            "x-label": {
              "en": "Organization",
              "pt-BR": "Organização"
            },
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Organization id for tenant users (admin/user)",
            "x-relation": {
              "entity": "Organization",
              "match": "id",
              "display": "name",
              "kind": "belongsTo"
            }
          },
          "password": {
            "x-label": {
              "en": "Password",
              "pt-BR": "Senha"
            },
            "type": "string",
            "minLength": 8,
            "format": "password",
            "description": "Password to access the system"
          },
          "emails": {
            "x-label": {
              "en": "Emails",
              "pt-BR": "E-mails"
            },
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Email"
            },
            "description": "User's e-mail addresses",
            "minItems": 1
          },
          "documents": {
            "x-label": {
              "en": "Documents",
              "pt-BR": "Documentos"
            },
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Document"
            },
            "description": "User's attached documents"
          },
          "phones": {
            "x-label": {
              "en": "Phones",
              "pt-BR": "Telefones"
            },
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Phone"
            },
            "description": "User's attached phone numbers"
          },
          "roles": {
            "x-label": {
              "en": "Roles",
              "pt-BR": "Papéis"
            },
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "User's access roles"
          },
          "createdAt": {
            "x-label": {
              "en": "Created at",
              "pt-BR": "Criado em"
            },
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "x-label": {
              "en": "Updated at",
              "pt-BR": "Atualizado em"
            },
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "x-label": {
              "en": "Deleted at",
              "pt-BR": "Excluído em"
            },
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "readOnly": true,
            "description": "Tombstone timestamp. Null while the record is alive."
          }
        }
      },
      "Organization": {
        "description": "Port output object for Organization resource.",
        "x-primary-key": "id",
        "x-service": "core",
        "required": ["id", "createdAt", "updatedAt", "name"],
        "type": "object",
        "properties": {
          "id": {
            "x-label": {
              "en": "ID",
              "pt-BR": "ID"
            },
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "x-label": {
              "en": "Name",
              "pt-BR": "Nome"
            },
            "type": "string"
          },
          "address": {
            "x-label": {
              "en": "Addresses",
              "pt-BR": "Endereços"
            },
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Address"
            }
          },
          "phone": {
            "x-label": {
              "en": "Phones",
              "pt-BR": "Telefones"
            },
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Phone"
            }
          },
          "email": {
            "x-label": {
              "en": "Email",
              "pt-BR": "E-mail"
            },
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Email"
            }
          },
          "users": {
            "x-label": {
              "en": "Members",
              "pt-BR": "Membros"
            },
            "type": "array",
            "description": "Member user ids.",
            "x-relation": {
              "entity": "User",
              "match": "id",
              "display": "username",
              "kind": "hasMany"
            },
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "x-label": {
              "en": "Created at",
              "pt-BR": "Criado em"
            },
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "x-label": {
              "en": "Updated at",
              "pt-BR": "Atualizado em"
            },
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "x-label": {
              "en": "Deleted at",
              "pt-BR": "Excluído em"
            },
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "readOnly": true,
            "description": "Tombstone timestamp. Null while the record is alive."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}